NIFI-4139 Moved key provider interface and implementations from nifi-data-provenance-utils module to nifi-security-utils module.
