+++ nifi-commons/nifi-data-provenance-utils/src/main/java/org/apache/nifi/provenance/ProvenanceEventEncryptor.java
@@ -17,6 +17,7 @@
 package org.apache.nifi.provenance;
 
 import java.security.KeyManagementException;
+import org.apache.nifi.security.kms.KeyProvider;
 
 public interface ProvenanceEventEncryptor {
